fix: warning on MacOS High Sierra and above related for forking
" +[__NSCFConstantString initialize] may have been in progress in
another thread when fork() was called. We cannot safely call it or
ignore it in the fork() child process. Crashing instead. Set a
breakpoint on objc_initializeAfterForkError to debug."

This is fixed by setting the multiprocess method to spawn instead
of fork.